Introducing Organizations
Ok please bear with me here, this is a spike into tryng to get organizations in
the picture. There are big organizations with LOTS of activity that make the
activity graph look sad.

This is because of the 300 events limitation of Github API. An improvement is,
for certain organizations, make the effor of going through Github Archive.

Organizations are basically the same as a user with slight differences:
* Organizations have members (not captured yet)
* Organizations are not hireable
* Right now we are not showing interesting projects

For all the rest, they are the same, if you check Recruiter Gem, you can get a
feeling of how closely related they are.